refactor: CustomListTest to ListImplementationTest
